Move items used in multiple places to common
