+++ src/server/common/helpers/build-running-services-row-headings.js
@@ -1,0 +1,12 @@
+import { kebabCase, upperFirst } from 'lodash'
+
+function buildRunningServicesRowHeadings(environments, size) {
+  return [
+    ...Object.keys(environments).map((key) => ({
+      text: upperFirst(kebabCase(key)),
+      size: 'small'
+    }))
+  ]
+}
+
+export { buildRunningServicesRowHeadings }
