+++ src/server/services/controllers/service.js
@@ -2,24 +2,14 @@ import Joi from 'joi'
 import Boom from '@hapi/boom'
 import { compose } from 'lodash/fp'
 
-import { config } from '~/src/config'
 import { sortByEnv } from '~/src/server/common/helpers/sort/sort-by-env'
 import { provideService } from '~/src/server/services/helpers/pre/provide-service'
 import { withEnvironments } from '~/src/server/common/transformers/with-environments'
 import { serviceToEntityDataList } from '~/src/server/services/transformers/service-to-entity-data-list'
-import { fetchRunningServicesById } from '~/src/server/services/helpers/fetch/fetch-running-services-by-id'
-import { runningServicesToEntityRow } from '~/src/server/services/transformers/running-services-to-entity-row'
-import { getEnvironments } from '~/src/server/common/helpers/environments/get-environments'
-import { kebabCase, upperFirst } from 'lodash'
-
-function buildRunningServicesRowHeadings(environments) {
-  return [
-    ...Object.keys(environments).map((key) => ({
-      text: upperFirst(kebabCase(key)),
-      size: 'medium'
-    }))
-  ]
-}
+import { fetchRunningServicesById } from '~/src/server/common/helpers/fetch/fetch-running-services-by-id'
+import { buildRunningServicesRowHeadings } from '~/src/server/common/helpers/build-running-services-row-headings'
+import { getEnvironmentsByTeam } from '~/src/server/common/helpers/environments/get-environments-by-team'
+import { runningServicesToEntityRow } from '~/src/server/common/transformers/running-services-to-entity-row'
 
 const serviceController = {
   options: {
@@ -35,10 +25,7 @@ const serviceController = {
     const serviceId = request.params?.serviceId
     const service = request.pre.service
 
-    const isPlatformService = service.teams
-      .map((team) => team.teamId)
-      .includes(config.get('oidcAdminGroupId'))
-    const environments = getEnvironments(isPlatformService)
+    const environments = getEnvironmentsByTeam(service.teams)
     const runningServices = await fetchRunningServicesById(serviceId)
     const runningServicesEntityRows = compose(
       runningServicesToEntityRow(environments),
