+++ src/server/services/helpers/pre/provide-service.test.js
@@ -1,11 +1,11 @@
 import { provideService } from '~/src/server/services/helpers/pre/provide-service'
 import { fetchRepository } from '~/src/server/services/helpers/fetch/fetch-repository'
-import { fetchDeployableService } from '~/src/server/services/helpers/fetch/fetch-deployable-service'
+import { fetchDeployableService } from '~/src/server/common/helpers/fetch/fetch-deployable-service'
 import { serviceDeployableFixture } from '~/src/__fixtures__/service-deployable'
 import { repositoryFixture } from '~/src/__fixtures__/repository'
 
 jest.mock('~/src/server/services/helpers/fetch/fetch-repository')
-jest.mock('~/src/server/services/helpers/fetch/fetch-deployable-service')
+jest.mock('~/src/server/common/helpers/fetch/fetch-deployable-service')
 
 describe('#provideService', () => {
   const mockRequest = {
